Refactor for kode-gjenbruk
