+++ src/server/api-handlers/driftsmeldinger.ts
@@ -1,25 +1,13 @@
 import fetch from 'node-fetch';
 import { RequestHandler } from 'express';
 import NodeCache from 'node-cache';
-import { tenSeconds } from '../utils';
+import { cachedResourceHandler } from './_cachedResourceHandler';
 
 const driftsmeldingerServiceUrl = `${process.env.API_XP_SERVICES_URL}/no.nav.navno/driftsmeldinger`;
 
 const cacheKey = 'driftsmeldinger-cache';
 
-const cache = new NodeCache({
-    stdTTL: tenSeconds,
-    deleteOnExpire: false,
-});
-
-let isFetching = false;
-
-const refreshCache = () => {
-    if (isFetching) {
-        return;
-    }
-
-    isFetching = true;
+const revalidateDriftsmeldingerCache = (cache: NodeCache) =>
     fetch(driftsmeldingerServiceUrl)
         .then((response) => {
             if (response.status === 200) {
@@ -36,23 +24,9 @@ const refreshCache = () => {
             console.error(`Failed to fetch from driftsmeldinger service - ${e}`);
             const prevCache = cache.get(cacheKey) || [];
             cache.set(cacheKey, prevCache);
-        })
-        .finally(() => {
-            isFetching = false;
         });
-};
 
-cache.on('expired', refreshCache);
-
-export const getDriftsmeldingerHandler: RequestHandler = (req, res) => {
-    const cached = cache.get(cacheKey);
-
-    if (cached) {
-        res.status(200).send(cached);
-    } else {
-        refreshCache();
-        cache.on('set', (key, value) => {
-            res.status(200).send(value);
-        });
-    }
-};
+export const getDriftsmeldingerHandler: RequestHandler = cachedResourceHandler(
+    revalidateDriftsmeldingerCache,
+    cacheKey
+);
